refactor: update order app

+++ services/order/src/main/kotlin/com/codejsha/bookstore/order/domain/handler/OrderItemHandler.kt
@@ -1,10 +1,13 @@
 package com.codejsha.bookstore.order.domain.handler
 
+import com.codejsha.bookstore.order.application.port.repo.OrderItemExtendedRepo
 import com.codejsha.bookstore.order.application.port.repo.OrderItemRepo
 import com.codejsha.bookstore.order.application.usecase.OrderItemCommand
 import com.codejsha.bookstore.order.application.usecase.OrderRead
 import com.codejsha.bookstore.order.domain.aggregate.entity.OrderItemEntity
 import com.codejsha.bookstore.order.domain.model.OrderItemDto
+
+import io.opentelemetry.instrumentation.annotations.WithSpan
 import org.reactivestreams.Publisher
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
@@ -12,30 +15,35 @@ import reactor.core.publisher.Flux
 
 interface OrderItemHandler {
     fun handle(command: OrderItemCommand): Flux<OrderItemEntity>
+
     fun handle(read: OrderRead): Publisher<OrderItemEntity>
 }
 
 @Component
 class OrderItemHandlerImpl(
-    private val orderItemRepo: OrderItemRepo
+    private val orderItemRepo: OrderItemRepo,
+    private val orderItemExtendedRepo: OrderItemExtendedRepo
 ) : OrderItemHandler {
-
     @Transactional
-    override fun handle(command: OrderItemCommand): Flux<OrderItemEntity> {
-        return when (command) {
+    @WithSpan
+    override fun handle(command: OrderItemCommand): Flux<OrderItemEntity> =
+        when (command) {
             is OrderItemCommand.CreateOrderItemsCommand -> {
-                val entities = command.newEntities()
+                val entities: List<OrderItemEntity> = command.newEntities()
                 orderItemRepo.saveAll(entities)
             }
 
             is OrderItemCommand.UpdateOrderItemsCommand -> {
-                orderItemRepo.findAllByOrderId(command.orderId)
+                orderItemExtendedRepo
+                    .findAllByOrderId(command.orderId)
                     .collectList()
                     .flatMapMany { existingEntities ->
-                        val synced = syncOrderItems(existingEntities, command.orderItems, command.orderId)
-                        val toDelete = existingEntities.filterNot { old ->
-                            synced.any { it.bookId == old.bookId }
-                        }
+                        val synced: List<OrderItemEntity> =
+                            syncOrderItems(existingEntities, command.orderItems, command.orderId)
+                        val toDelete: List<OrderItemEntity?> =
+                            existingEntities.filterNot { old ->
+                                synced.any { it.bookId == old.bookId }
+                            }
 
                         val saveFlux = orderItemRepo.saveAll(synced)
                         val deleteMono = orderItemRepo.deleteAll(toDelete).then()
@@ -44,11 +52,11 @@ class OrderItemHandlerImpl(
             }
 
             is OrderItemCommand.DeleteOrderItemsCommand -> {
-                orderItemRepo.deleteById(command.id)
+                orderItemRepo
+                    .deleteById(command.id)
                     .thenMany(Flux.empty())
             }
         }
-    }
 
     fun syncOrderItems(
         existingEntities: List<OrderItemEntity>,
@@ -60,16 +68,17 @@ class OrderItemHandlerImpl(
 
         for (dto in incomingDtos) {
             val existing = existingMap.remove(dto.bookId)
-            val entity = existing?.apply {
-                if (quantity != dto.quantity) {
+            val entity =
+                existing?.apply {
+                    if (quantity != dto.quantity) {
+                        quantity = dto.quantity
+                    }
+                } ?: OrderItemEntity(
+                    id = null,
+                    orderId = orderId,
+                    bookId = dto.bookId,
                     quantity = dto.quantity
-                }
-            } ?: OrderItemEntity(
-                id = null,
-                orderId = orderId,
-                bookId = dto.bookId,
-                quantity = dto.quantity
-            )
+                )
             updatedEntities += entity
         }
 
@@ -77,8 +86,9 @@ class OrderItemHandlerImpl(
     }
 
     @Transactional(readOnly = true)
-    override fun handle(read: OrderRead): Publisher<OrderItemEntity> {
-        return when (read) {
+    @WithSpan
+    override fun handle(read: OrderRead): Publisher<OrderItemEntity> =
+        when (read) {
             is OrderRead.FindAllOrdersRead -> {
                 orderItemRepo.findAll()
             }
@@ -87,5 +97,4 @@ class OrderItemHandlerImpl(
                 orderItemRepo.findById(read.id)
             }
         }
-    }
 }
